Verifications are now fell into VerificationManager
