+++ src/manager/manager.cpp
@@ -35,6 +35,9 @@ namespace qls
             // sql更新初始化数据
             // ...
         }
+
+        m_dataManager.init();
+        m_verificationManager.init();
     }
 
     void Manager::addUserSocket2GlobalRoom(long long user_id, const std::shared_ptr<Socket>& socket_ptr)
@@ -230,309 +233,18 @@ namespace qls
         return m_user_map;
     }
 
-    void Manager::addFriendRoomVerification(long long user_id_1, long long user_id_2)
-    {
-        {
-            std::unique_lock<std::shared_mutex> ul(m_FriendRoomVerification_map_mutex);
-
-            if (m_FriendRoomVerification_map.find({ user_id_1, user_id_2 }) ==
-                m_FriendRoomVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            m_FriendRoomVerification_map.emplace(PrivateRoomIDStruct{ user_id_1, user_id_2 },
-                                                 FriendRoomVerification{ user_id_1, user_id_2 });
-        }
-        
-        // user1
-        {
-            qls::User::UserVerificationStruct uv;
-
-            uv.user_id = user_id_2;
-            uv.verification_type =
-                qls::User::UserVerificationStruct::VerificationType::Sent;
-
-            auto ptr = this->getUser(user_id_1);
-            ptr->addFriendVerification(user_id_2, std::move(uv));
-
-            // 未完成 通知另一方
-            // m_globalRoom->baseSendData()
-        }
-
-        // user2
-        {
-            qls::User::UserVerificationStruct uv;
-
-            uv.user_id = user_id_1;
-            uv.verification_type =
-                qls::User::UserVerificationStruct::VerificationType::Received;
-
-            auto ptr = this->getUser(user_id_2);
-            ptr->addFriendVerification(user_id_1, std::move(uv));
-
-            // 未完成 通知另一方
-            // m_globalRoom->baseSendData()
-        }
-    }
-
-    bool Manager::hasFriendRoomVerification(long long user_id_1, long long user_id_2) const
-    {
-        std::shared_lock<std::shared_mutex> sl(m_FriendRoomVerification_map_mutex);
-        
-        return m_FriendRoomVerification_map.find({ user_id_1, user_id_2 }) !=
-            m_FriendRoomVerification_map.end();
-    }
-
-    bool Manager::setFriendVerified(long long user_id_1, long long user_id_2,
-                                    long long user_id, bool is_verified)
-    {
-        bool result = false;
-        {
-            std::unique_lock<std::shared_mutex> ul(m_FriendRoomVerification_map_mutex);
-
-            auto itor = m_FriendRoomVerification_map.find({ user_id_1, user_id_2 });
-            if (itor == m_FriendRoomVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            auto& ver = itor->second;
-            ver.setUserVerified(user_id, is_verified);
-
-            result = ver.getUserVerified(user_id_1) && ver.getUserVerified(user_id_2);
-        }
-
-        if (result)
-        {
-            bool error = false;
-            try
-            {
-                this->getPrivateRoomId(user_id_1, user_id_2);
-                error = true;
-            }
-            catch (...)
-            {
-                std::unique_lock<std::shared_mutex> ul(m_basePrivateRoom_map_mutex);
-                this->addPrivateRoom(user_id_1, user_id_2);
-
-                // 更新user1的friendList
-                {
-                    std::shared_ptr<qls::User> ptr;
-                    {
-                        std::shared_lock<std::shared_mutex> sl(m_user_map_mutex);
-                        auto itor = m_user_map.find(user_id_1);
-                        if (itor == m_user_map.end())
-                            throw std::invalid_argument("Wrong argument!");
-                        ptr = itor->second;
-                    }
-                    auto set = std::move(ptr->getFriendList());
-                    set.insert(user_id_2);
-                    ptr->updateFriendList(std::move(set));
-
-                    ptr->removeFriendVerification(user_id_2);
-                }
-                
-                // 更新user2的friendList
-                {
-                    std::shared_ptr<qls::User> ptr;
-                    {
-                        std::shared_lock<std::shared_mutex> sl(m_user_map_mutex);
-                        auto itor = m_user_map.find(user_id_2);
-                        if (itor == m_user_map.end())
-                            throw std::invalid_argument("Wrong argument!");
-                        ptr = itor->second;
-                    }
-                    auto set = std::move(ptr->getFriendList());
-                    set.insert(user_id_1);
-                    ptr->updateFriendList(std::move(set));
-
-                    ptr->removeFriendVerification(user_id_1);
-                }
-
-                // 未完成 通知另一方
-                // m_globalRoom->baseSendData()
-            }
-
-            if (error) throw std::invalid_argument("Wrong argument!");
-        }
-
-        return result;
-    }
-
-    void Manager::removeFriendRoomVerification(long long user_id_1, long long user_id_2)
-    {
-        {
-            std::unique_lock<std::shared_mutex> ul(m_FriendRoomVerification_map_mutex);
-
-            auto itor = m_FriendRoomVerification_map.find({ user_id_1, user_id_2 });
-            if (itor == m_FriendRoomVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            m_FriendRoomVerification_map.erase(itor);
-        }
-
-        this->getUser(user_id_1)->removeFriendVerification(user_id_2);
-        this->getUser(user_id_2)->removeFriendVerification(user_id_1);
-    }
-
-    void Manager::addGroupRoomVerification(long long group_id, long long user_id)
-    {
-        {
-            std::unique_lock<std::shared_mutex> ul(m_GroupVerification_map_mutex);
-
-            if (m_GroupVerification_map.find({ group_id, user_id }) ==
-                m_GroupVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            m_GroupVerification_map.emplace(GroupVerificationStruct{ group_id, user_id },
-                                            GroupRoomVerification{ group_id, user_id });
-        }
-
-        // 用户发送请求
-        {
-            qls::User::UserVerificationStruct uv;
-
-            uv.user_id = group_id;
-            uv.verification_type =
-                qls::User::UserVerificationStruct::VerificationType::Sent;
-
-            auto ptr = this->getUser(user_id);
-            ptr->addGroupVerification(group_id, std::move(uv));
-        }
-
-        // 群聊拥有者接收请求
-        {
-            qls::User::UserVerificationStruct uv;
-
-            uv.user_id = user_id;
-            uv.verification_type =
-                qls::User::UserVerificationStruct::VerificationType::Received;
-
-            auto ptr = this->getUser(this->getGroupRoom(group_id)->getAdministrator());
-            ptr->addGroupVerification(group_id, std::move(uv));
-        }
-
-        // 未完成 通知另一方
-        // m_globalRoom->baseSendData()
-    }
-
-    bool Manager::hasGroupRoomVerification(long long group_id, long long user_id) const
-    {
-        std::shared_lock<std::shared_mutex> sl(m_GroupVerification_map_mutex);
-
-        return m_GroupVerification_map.find({ group_id, user_id }) !=
-            m_GroupVerification_map.end();
-    }
-
-    bool Manager::setGroupRoomGroupVerified(long long group_id, long long user_id, bool is_verified)
-    {
-        bool result = false;
-        {
-            std::unique_lock<std::shared_mutex> ul(m_GroupVerification_map_mutex);
-
-            auto itor = m_GroupVerification_map.find({ group_id, user_id });
-            if (itor == m_GroupVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            auto& ver = itor->second;
-            ver.setGroupVerified(is_verified);
-            result = ver.getGroupVerified() && ver.getUserVerified();
-        }
-
-        if (result)
-        {
-            std::shared_lock<std::shared_mutex> sl(m_baseRoom_map_mutex);
-
-            auto itor = m_baseRoom_map.find(group_id);
-            if (itor == m_baseRoom_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            itor->second->addMember(user_id);
-
-            // 更新user的groupList
-            std::shared_ptr<qls::User> ptr;
-            {
-                std::shared_lock<std::shared_mutex> sl(m_user_map_mutex);
-                auto itor = m_user_map.find(user_id);
-                if (itor == m_user_map.end())
-                    throw std::invalid_argument("Wrong argument!");
-                ptr = itor->second;
-            }
-            auto set = std::move(ptr->getGroupList());
-            set.insert(group_id);
-            ptr->updateGroupList(std::move(set));
-
-            // 未完成 通知另一方
-            // m_globalRoom->baseSendData()
-        }
-        return result;
-    }
-
-    bool Manager::setGroupRoomUserVerified(long long group_id, long long user_id, bool is_verified)
-    {
-        bool result = false;
-        {
-            std::unique_lock<std::shared_mutex> ul(m_GroupVerification_map_mutex);
-
-            auto itor = m_GroupVerification_map.find({ group_id, user_id });
-            if (itor == m_GroupVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            auto& ver = itor->second;
-            ver.setUserVerified(is_verified);
-            result = ver.getGroupVerified() && ver.getUserVerified();
-        }
-
-        if (result)
-        {
-            std::shared_lock<std::shared_mutex> sl(m_baseRoom_map_mutex);
-
-            auto itor = m_baseRoom_map.find(group_id);
-            if (itor == m_baseRoom_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            itor->second->addMember(user_id);
-
-            // 更新user的friendlist
-            std::shared_ptr<qls::User> ptr;
-            {
-                std::shared_lock<std::shared_mutex> sl(m_user_map_mutex);
-                auto itor = m_user_map.find(user_id);
-                if (itor == m_user_map.end())
-                    throw std::invalid_argument("Wrong argument!");
-                ptr = itor->second;
-            }
-            auto set = std::move(ptr->getGroupList());
-            set.insert(group_id);
-            ptr->updateGroupList(std::move(set));
-
-            ptr->removeGroupVerification(group_id, user_id);
-
-            {
-                auto ptr = this->getUser(this->getGroupRoom(group_id)->getAdministrator());
-                ptr->removeGroupVerification(group_id, user_id);
-            }
-        }
-        return result;
-    }
-
-    void Manager::removeGroupRoomVerification(long long group_id, long long user_id)
-    {
-        {
-            std::unique_lock<std::shared_mutex> ul(m_GroupVerification_map_mutex);
-
-            auto itor = m_GroupVerification_map.find({ group_id, user_id });
-            if (itor == m_GroupVerification_map.end())
-                throw std::invalid_argument("Wrong argument!");
-
-            m_GroupVerification_map.erase(itor);
-        }
-
-        this->getUser(getGroupRoom(group_id)->getAdministrator())
-            ->removeGroupVerification(group_id, user_id);
-
-        this->getUser(user_id)->removeGroupVerification(group_id, user_id);
-    }
-
     quqisql::SQLDBProcess& Manager::getServerSqlProcess()
     {
         return this->m_sqlProcess;
+    }
+
+    qls::DataManager &Manager::getServerDataManager()
+    {
+        return m_dataManager;
+    }
+
+    qls::VerificationManager &Manager::getServerVerificationManager()
+    {
+        return m_verificationManager;
     }
 }
